@@ -1,9 +1,8 @@
-﻿namespace FluentTimeSeries
-{
-    public class GaussianRandomConfiguration
-    {
-        public double Mean { get; set; }
+﻿namespace FluentTimeSeries;
 
-        public double StdDev { get; set; } = 1.0;
-    }
+public class GaussianRandomConfiguration
+{
+    public double Mean { get; set; }
+
+    public double StdDev { get; set; } = 1.0;
 }
