@@ -1,13 +1,22 @@
 ﻿using System;
 
-namespace FluentTimeSeries
+namespace FluentTimeSeries;
+
+public class TimeSeriesOrigin
 {
-    public interface ITimeSeries
+    public DateTime Origin => DateTime.Now;
+
+    public TimeSeriesOrigin()
     {
-        DataPoint Next(DateTime? timestamp = null);
-
-        DataPoint[] Block(int numOfPoints, double samplingInterval, DateTime? origin);
-
-        DataPoint[] Block(TimeSpan length, double samplingInterval, DateTime? origin);
+        
     }
+}
+
+public interface ITimeSeries
+{
+    DataPoint[] Block(int numOfPoints, double samplingInterval, DateTime? origin);
+
+    DataPoint[] Block(TimeSpan length, double samplingInterval, DateTime? origin);
+
+    DataPoint Sample(DateTime? timestamp = null);
 }
