@@ -1,9 +1,8 @@
-﻿namespace FluentTimeSeries
-{
-    public class UniformRandomConfiguration
-    {
-        public double Max { get; set; } = 1.0;
+﻿namespace FluentTimeSeries;
 
-        public double Min { get; set; }
-    }
+public class UniformRandomConfiguration
+{
+    public double Max { get; set; } = 1.0;
+
+    public double Min { get; set; }
 }
